chore: Wrapper should not explicitly handle Feature Flag API
!!BREAKING CHANGE!!

Feature flags are now defined in their own object where the object key
is the literal FF string found in the [SDK
docs](https://jitsi.github.io/handbook/docs/dev-guide/mobile-feature-flags),
 and the value is a boolean.

Example:

```
const options = {
   room: ... ,
   userInfo: ... ,
   featureFlags: {
      "live-streaming.enabled": false
   }
}
```

There is no need for this wrapper to be a middleman between the consumer
and the SDK when setting feature flags. This only adds more work on
maintaining the API since everytime the SDK changes their feature flags
API, we have to make the corresponding change on the wrapper.